Extract binding module type writing logic into separate class
